% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{SchemaForm}
\alias{SchemaForm}
\title{SchemaForm}
\usage{
SchemaForm(inputId, schema, uiSchema = NULL, formData = NULL, ...)
}
\arguments{
\item{inputId}{The inputId as used by observeEvents or outputs (input[\link{inputId}])

Within your observeEvents/reactives/outputs, you've got access to two inputs created by this form.
\itemize{
\item input$\code{inputId}' which gets triggered on every form update.
\item 'input$\code{inputId}_submit' when the submit button is pressed.
}

In both inputs, the formData is captured, which can be manipulated directly or converted back to JSON using \code{jsonlite::toJSON}}

\item{schema}{The JSON schema on which to base the form. This can either be:
\itemize{
\item An imported JSON file (the resulting variable of a call to \code{jsonlite::fromJSON()})
\item The path to a JSON file (\code{"./path/to/file.json"})
\item A raw JSON string (e.g. \verb{\{ title: 'Test form', type: 'string', \}})
}}

\item{uiSchema}{The uiSchema is used to add more customization to the form's look and feel.

The same input formats as for \code{schema} are possible (import JSON, JSON file path, JSON string)}

\item{formData}{A JSON used to prefill the form.

The same input formats as for \code{schema} are possible (import JSON, JSON file path, JSON string)}

\item{...}{Additional props to pass on to the Form React component.}
}
\description{
Create a React Json Schema Form within an R Shiny app (https://rjsf-team.github.io/react-jsonschema-form/docs/)
}
